LayoutTests/fast/forms/radio/radio-group-in-detached-form.html

<!DOCTYPE html>
<script src="../../js/resources/js-test-pre.js"></script>
<script>
description('Radio buttons within the same form outside a document should still be updated appropriately when one of them is checked.');

var form = document.createElement('form');
var radio1 = form.appendChild(document.createElement('input'));
var radio2 = form.appendChild(document.createElement('input'));
radio1.type = radio2.type = 'radio';
radio1.name = radio2.name = 'pref';
radio1.checked = true;
radio2.checked = true;
shouldBeFalse('radio1.checked');
</script>
<script src="../../js/resources/js-test-post.js"></script>
